Simplify cashflow data structure and refactor fromtend data processing
Cashflow data is being reduced to one object to reduce data redundancy. Also, the CashflowGraph requires an
object to parse, as does the Cashflow form. Therefore that extra function has been abstracted from CashflowGraph
because there are multiple places where a string/file contents must be parsed. Cumulative revenue over time
is also calculated for the graph in this way